+++ internal/services/app.go
@@ -247,102 +247,3 @@ func (s *AppService) BuildApp() {
 	// Fill the table with the initial data
 	s.fillTable(s.packages)
 }
-
-func (s *AppService) handleKeyEventInput(event *tcell.EventKey) *tcell.EventKey {
-	if s.LayoutService.GetSearchField().HasFocus() {
-		return event
-	}
-
-	switch event.Key() {
-	case tcell.KeyRune:
-		switch event.Rune() {
-		case 'q': // Quit the application
-			s.app.Stop()
-			return nil
-		case 'u': // Update the selected package
-			row, _ := s.LayoutService.GetTableResult().GetSelection()
-			if row > 0 {
-				info := (*s.filteredPackages)[row-1]
-				modal := s.LayoutService.GenerateModal(fmt.Sprintf("Are you sure you want to update the package: %s?", info.Name), func() {
-					s.app.SetRoot(s.LayoutService.GetGrid(), true).SetFocus(s.LayoutService.GetTableResult())
-					s.LayoutService.GetOutputView().Clear()
-					go func() {
-						_ = s.CommandService.UpdatePackage(info, s.app, s.LayoutService.GetOutputView())
-						s.updateTableView()
-					}()
-				}, func() {
-					s.app.SetRoot(s.LayoutService.GetGrid(), true).SetFocus(s.LayoutService.GetTableResult())
-				})
-
-				s.app.SetRoot(modal, true).SetFocus(modal)
-			}
-			return nil
-		case 'r': // Remove the selected package
-			row, _ := s.LayoutService.GetTableResult().GetSelection()
-			if row > 0 {
-				info := (*s.filteredPackages)[row-1]
-				modal := s.LayoutService.GenerateModal(fmt.Sprintf("Are you sure you want to remove the package: %s?", info.Name), func() {
-					s.app.SetRoot(s.LayoutService.GetGrid(), true).SetFocus(s.LayoutService.GetTableResult())
-					s.LayoutService.GetOutputView().Clear()
-					go func() {
-						_ = s.CommandService.RemovePackage(info, s.app, s.LayoutService.GetOutputView())
-						s.updateTableView()
-					}()
-				}, func() {
-					s.app.SetRoot(s.LayoutService.GetGrid(), true).SetFocus(s.LayoutService.GetTableResult())
-				})
-				s.app.SetRoot(modal, true).SetFocus(modal)
-			}
-			return nil
-		case 'i': // Install the selected package
-			row, _ := s.LayoutService.GetTableResult().GetSelection()
-			if row > 0 {
-				info := (*s.filteredPackages)[row-1]
-				modal := s.LayoutService.GenerateModal(fmt.Sprintf("Are you sure you want to install the package: %s?", info.Name), func() {
-					s.app.SetRoot(s.LayoutService.GetGrid(), true).SetFocus(s.LayoutService.GetTableResult())
-					s.LayoutService.GetOutputView().Clear()
-					go func() {
-						_ = s.CommandService.InstallPackage(info, s.app, s.LayoutService.GetOutputView())
-						s.updateTableView()
-					}()
-				}, func() {
-					s.app.SetRoot(s.LayoutService.GetGrid(), true).SetFocus(s.LayoutService.GetTableResult())
-				})
-				s.app.SetRoot(modal, true).SetFocus(modal)
-			}
-			return nil
-		case '/':
-			s.app.SetFocus(s.LayoutService.GetSearchField())
-			return nil
-		case 'f':
-			s.showOnlyInstalled = !s.showOnlyInstalled
-			if s.showOnlyInstalled {
-				s.LayoutService.GetSearchField().SetLabel("Search (Installed): ")
-			} else {
-				s.LayoutService.GetSearchField().SetLabel("Search (All): ")
-			}
-			s.applySearchFilter(s.LayoutService.GetSearchField().GetText())
-			s.LayoutService.GetTableResult().ScrollToBeginning()
-			return nil
-		}
-	case tcell.KeyCtrlU:
-		// Update homebrew
-		modal := s.LayoutService.GenerateModal("Are you sure you want to update Homebrew?", func() {
-			s.app.SetRoot(s.LayoutService.GetGrid(), true).SetFocus(s.LayoutService.GetTableResult())
-			s.LayoutService.GetOutputView().Clear()
-			go func() {
-				_ = s.CommandService.UpdateHomebrew(s.app, s.LayoutService.GetOutputView())
-				s.updateTableView()
-			}()
-		}, func() {
-			s.app.SetRoot(s.LayoutService.GetGrid(), true).SetFocus(s.LayoutService.GetTableResult())
-		})
-		s.app.SetRoot(modal, true).SetFocus(modal)
-		return nil
-	case tcell.KeyEsc:
-		s.app.SetRoot(s.LayoutService.GetGrid(), true).SetFocus(s.LayoutService.GetTableResult())
-		return nil
-	}
-
-	return event
-}
